+++ CHANGELOG.rst
@@ -6,6 +6,7 @@
 * `repr()` will print matrix images as matrices [jcupitt]
 * more robust bandwise index/slice; added fancy slicing (step != 1) [erdmann]
 * fig segfault for `im.bandjoin([])`.  Now returns `im` [erdmann]
+* add numpy-style extended indexing (index with list of ints or bools) [erdmann]
 
 ## Version 2.1.16 (28 Jun 2021)